% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggplot2_theme_to_plotly}
\alias{ggplot2_theme_to_plotly}
\title{Convert ggplot2 theme to plotly layout}
\usage{
ggplot2_theme_to_plotly(theme)
}
\arguments{
\item{theme}{A ggplot2 theme object}
}
\value{
A list of plotly layout settings
}
\description{
Helper function to convert ggplot2 theme settings to plotly layout
}
\examples{
\dontrun{
theme_settings <- ggplot2_theme_to_plotly(theme_cq_html())
}
}
