% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_plot}
\alias{save_plot}
\title{Save a ggplot with consistent settings}
\usage{
save_plot(plot, filename, width = 8, height = 6, dpi = 300, device = "png")
}
\arguments{
\item{plot}{A ggplot object to save}

\item{filename}{Filename to save the plot}

\item{width}{Width of the plot in inches}

\item{height}{Height of the plot in inches}

\item{dpi}{Resolution in dots per inch}

\item{device}{Device to use, defaults to "png"}
}
\value{
Invisibly returns the path to the saved file
}
\description{
Save a ggplot with consistent settings
}
\examples{
\dontrun{
p <- ggplot(mtcars, aes(x = mpg, y = hp)) + geom_point() + theme_cq_pdf()
save_plot(p, "myplot.png", width = 8, height = 6)
}
}
